+++ Sources/Pathos/copy.swift
@@ -5,8 +5,8 @@ import Darwin
 #endif
 
 // TODO: missing docstring.
-/// - SeeAlso: To work with `Path` or `PathRepresentable`, use `PathRepresentable.copy(to:followSymbolicLink:chunkSize:)`.
-public func copyFile(fromPath source: String, toPath destination: String, followSymbolicLink: Bool = true, chunkSize: Int = 1024 * 16) throws {
+/// - SeeAlso: To work with `Path` or `PathRepresentable`, use `PathRepresentable.copy(to:followSymlink:chunkSize:)`.
+public func copyFile(fromPath source: String, toPath destination: String, followSymlink: Bool = true, chunkSize: Int = 1024 * 16) throws {
     var sourceStatus = try _lstat(at: source)
     let destinationStatus = try? _stat(at: destination)
     if _ifmt(sourceStatus) != S_IFREG && _ifmt(sourceStatus) != S_IFLNK {
@@ -24,7 +24,7 @@ public func copyFile(fromPath source: String, toPath destination: String, follow
     // some question from Python's standard library: What about other special files? (sockets, devices...)
 
     let isLink = _ifmt(sourceStatus) == S_IFLNK
-    if !followSymbolicLink && isLink {
+    if !followSymlink && isLink {
         try createSymbolicLink(fromPath: readSymbolicLink(atPath: source), toPath: destination)
         return
     }
@@ -64,11 +64,11 @@ public func copyFile(fromPath source: String, toPath destination: String, follow
 
 extension PathRepresentable {
     // TODO: missing docstring.
-    /// - SeeAlso: `copyFile(fromPath:toPath:followSymbolicLink:chunkSize:)`.
+    /// - SeeAlso: `copyFile(fromPath:toPath:followSymlink:chunkSize:)`.
     @discardableResult
-    public func copy(to destination: PathRepresentable, followSymbolicLink: Bool = true, chunkSize: Int = 1024 * 16) -> Bool {
+    public func copy(to destination: PathRepresentable, followSymlink: Bool = true, chunkSize: Int = 1024 * 16) -> Bool {
         do {
-            try copyFile(fromPath: self.pathString, toPath: destination.pathString, followSymbolicLink: followSymbolicLink, chunkSize: chunkSize)
+            try copyFile(fromPath: self.pathString, toPath: destination.pathString, followSymlink: followSymlink, chunkSize: chunkSize)
         } catch {
             return false
         }